requirements.txt
Django==5.1.4
django-allauth
certifi
Bootstrap5
APScheduler==3.11.0
pytest==8.3.5
pytest-django==4.10.0
django-vote==2.4.0
pillow==11.1.0
psycopg2-binary


